+++ src/dac_jax/model/discriminator.py
@@ -8,6 +8,7 @@ from jax import numpy as jnp
 
 from dac_jax.audio_utils import stft
 from dac_jax.nn.weight_norm import WeightNorm as MyWeightNorm
+from dac_jax.nn.layers import make_initializer
 
 
 class LeakyReLU(nn.Module):
@@ -19,19 +20,6 @@ class LeakyReLU(nn.Module):
         return nn.leaky_relu(x, negative_slope=self.negative_slope)
 
 
-def make_initializer(in_channels, out_channels, kernel_size, groups, mode="fan_in"):
-    # https://pytorch.org/docs/stable/generated/torch.nn.Conv1d.html
-    if mode == "fan_in":
-        c = in_channels
-    elif mode == "fan_out":
-        c = out_channels
-    else:
-        raise ValueError(f"Unexpected mode: {mode}")
-    k = groups / (c * jnp.prod(jnp.array(kernel_size)))
-    scale = jnp.sqrt(k)
-    return lambda key, shape, dtype: jax.random.uniform(key, shape, minval=-scale, maxval=scale, dtype=dtype)
-
-
 class WNConv(nn.Conv):
 
     act: bool = True
@@ -40,13 +28,13 @@ class WNConv(nn.Conv):
     def __call__(self, x):
 
         kernel_init = make_initializer(
-            x.shape[-1], self.features, self.kernel_size, self.feature_group_count
+            x.shape[-1], self.features, self.kernel_size, self.feature_group_count, mode="fan_in",
         )
 
         if self.use_bias:
             # note: we just ignore whatever self.bias_init is
             bias_init = make_initializer(
-                x.shape[-1], self.features, self.kernel_size, self.feature_group_count
+                x.shape[-1], self.features, self.kernel_size, self.feature_group_count, mode="fan_in",
             )
         else:
             bias_init = None
